requirements.txt
Flask==1.0.2
PyYAML==3.13
monkey
gevent
requests